Mesh-type printing, simplify Wifi-handling
Print out the type of mesh (similar to as how we do it for WiFi as well).

Simplify the Wifi-handling more - return with 0 if the SSID or
password are too long. As we have now just one WiFi branch in
the connection we can do it with one if.
- No extra memory allocations/hard-coded buffers
- No copying, just pass the pointer or use the mbed_app -defines.

I considered doing a calloc() and then strncopy(), but that would
leave us with dangling pointers we could never free.