PIN defines in mbed_lib.json & README update
Change the way we handle the default pinouts - not via mbed_app.json,
but rather have them in mbed_lib.json. This is a non-backwards
compatible change to certain extent, as the current pin defines
potentially residing in mbed_app.json would not be valid anymore.

Potential overrides need to be done differently, as the define
one has to override is the easy-connect override,
not the app override.

Error handling was giving still wrong note in 2nd branch of
easy_connect(), fixed that (ref: review feedback from Brian Daniels).

README.md also updated a bit about the related changes.